Use generic function for status changes
