+++ modules/040-node-manager/hooks/util.go
@@ -21,7 +21,10 @@ import (
 	"fmt"
 
 	"github.com/Masterminds/semver/v3"
+	"github.com/deckhouse/deckhouse/modules/040-node-manager/hooks/internal/mcm/v1alpha1"
+	ngv1 "github.com/deckhouse/deckhouse/modules/040-node-manager/hooks/internal/v1"
 	"github.com/flant/addon-operator/sdk"
+	"github.com/flant/shell-operator/pkg/kube/object_patch"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
@@ -82,4 +85,56 @@ func semverMin(versions []*semver.Version) *semver.Version {
 		}
 	}
 	return res
+}
+
+func patchNodeGroupStatus(patcher *object_patch.PatchCollector, nodeGroupName string, patch interface{}) {
+	patcher.MergePatch(patch, "deckhouse.io/v1", "NodeGroup", "", nodeGroupName, object_patch.WithSubresource("/status"))
+}
+
+func setNodeGroupStatus(patcher *object_patch.PatchCollector, nodeGroupName string, statusField string, value interface{}) {
+	statusPatch := map[string]interface{}{
+		"status": map[string]interface{}{
+			statusField: value,
+		},
+	}
+	patchNodeGroupStatus(patcher, nodeGroupName, statusPatch)
+}
+
+func buildUpdateStatusPatch(
+	nodesNum, readyNodesNum, uptodateNodesCount, minPerZone, maxPerZone, desiredMax, instancesNum int32,
+	nodeType ngv1.NodeType, statusMsg string,
+	lastMachineFailures []*v1alpha1.MachineSummary,
+) interface{} {
+	ready := "True"
+	if len(statusMsg) > 0 {
+		ready = "False"
+	}
+
+	patch := map[string]interface{}{
+		"nodes":    nodesNum,
+		"ready":    readyNodesNum,
+		"upToDate": uptodateNodesCount,
+	}
+	if nodeType == ngv1.NodeTypeCloudEphemeral {
+		patch["min"] = minPerZone
+		patch["max"] = maxPerZone
+		patch["desired"] = desiredMax
+		patch["instances"] = instancesNum
+		patch["lastMachineFailures"] = lastMachineFailures
+
+		if len(lastMachineFailures) == 0 {
+			patch["lastMachineFailures"] = make([]interface{}, 0) // to make [] array in json result
+		}
+	}
+
+	patch["conditionSummary"] = map[string]interface{}{
+		"ready":         ready,
+		"statusMessage": statusMsg,
+	}
+
+	statusPatch := map[string]interface{}{
+		"status": patch,
+	}
+
+	return statusPatch
 }
